% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxplot-package.R
\docType{package}
\name{cxplot-package}
\alias{cxplot}
\alias{cxplot-package}
\title{cxplot: A Wrapper for CanvasXpress That Implements the Grammar of the R ggplot2 Package Written Completely in Javascript}
\description{
A companion package and wrapper for the JavaScript-based CanvasXpress, visualization library that allows use of ggplot2 grammar.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
